feat: refactor TranscriptionContext and implement custom hooks for provider settings and initialization
